+++ CHANGELOG.md
@@ -13,6 +13,7 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 ### Fixed
 
 - App menu label visibility compatibility issue with open whether extension.
+- Hidden workspace switcher padding in overview.
 - Keep clock menu position when another widget adds to the panel.
 - Looking glass API conflict with Dash to Panel extension.
 - Looking glass error when primary monitor doesn't exist.